moved non blocking inmplementation to tryLock method
